+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the **F_create_bufferify_function** option.
 - Support `true` and `false` with implied attribute.
 - Process 'void **' argument as 'type(C_PTR)' by reference.
+- Added assumedtype attribute to use Fortran assumed-type declaration,
+  `type(*)`, for `void *` arguments.
 
 ### Fixed
 - C++ function arguments which pass a class by value.